backend/requirements.txt
# Django Framework
Django>=5.0.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2>=2.9.9

# Environment Variables
python-decouple>=3.8
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# CORS Headers (for frontend integration)
django-cors-headers>=4.3.0
